Move to folder dim2.
